Create utils module with custom process lock function
