+++ install/uvm-install2/Cargo.toml
@@ -12,4 +12,3 @@ serde = "1.0"
 serde_derive = "1.0"
 uvm_cli = { path = "../../uvm_cli" }
 uvm_core = { path = "../../uvm_core" }
-cluFlock = "0.1.5"
