+++ uvm_core/src/lib.rs
@@ -24,6 +24,24 @@ extern crate dirs;
 #[macro_use]
 extern crate itertools;
 
+pub mod utils;
+
+#[macro_export]
+#[cfg(unix)]
+macro_rules! lock_process {
+
+    ($lock_path:expr) => (
+        let lock_file = fs::File::create($lock_path)?;
+        let _lock = utils::lock_process_or_wait(&lock_file)?;
+    )
+}
+
+#[macro_export]
+#[cfg(windows)]
+macro_rules! lock_process {
+    ($lock_path:expr) => ()
+}
+
 #[macro_export]
 macro_rules! cargo_version {
     // `()` indicates that the macro takes no argument.
